final commit

+++ lib/view/widgets/core_widgets.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:flutter/material.dart';
 import 'package:movielist/controller/movielist_provider.dart';
 import 'package:movielist/view/widgets/const_colors.dart';
@@ -36,39 +38,39 @@ class ButtonWidget extends StatelessWidget {
   }
 }
 
-
-Future<void> snackBar(BuildContext context, text,{color}) async {
+Future<void> snackBar(BuildContext context, text, {color}) async {
   ScaffoldMessenger.of(context).showSnackBar(SnackBar(
     content: Text(text),
-    backgroundColor:color?? kGreenColor,
+    backgroundColor: color ?? kGreenColor,
     behavior: SnackBarBehavior.floating,
     margin: const EdgeInsets.all(10),
   ));
 }
 
-  Future<void> deleteMovieButton(int index, context) async {
-    showDialog(
-        context: context,
-        builder: (context) {
-          return AlertDialog(
-            title: const Text('Are you sure want to delete ?'),
-            actions: [
-              TextButton(
+Future<void> deleteMovieButton(int? index, context) async {
+  log(index.toString());
+  showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('Are you sure want to delete ?'),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Provider.of<MovieListProvider>(context, listen: false)
+                    .deleteMovie(index);
+                Navigator.of(context).pop();
+                Navigator.of(context).pop();
+                snackBar(context, 'Succesfully deleted');
+              },
+              child: const Text('Yes'),
+            ),
+            TextButton(
                 onPressed: () {
-                  
-                  Provider.of<MovieListProvider>(context, listen: false)
-        .deleteMovie(index);
                   Navigator.of(context).pop();
-                  snackBar(context, 'Succesfully deleted');
                 },
-                child: const Text('Yes'),
-              ),
-              TextButton(
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                  },
-                  child: const Text('No'))
-            ],
-          );
-        });
-  }
+                child: const Text('No'))
+          ],
+        );
+      });
+}
